+++ lib/aws-sdk-rails.rb
@@ -6,9 +6,11 @@ require_relative 'aws/rails/notifications'
 require_relative 'aws/rails/sqs_active_job/configuration'
 require_relative 'aws/rails/sqs_active_job/executor'
 require_relative 'aws/rails/sqs_active_job/job_runner'
+require_relative 'aws/rails/sqs_active_job/lambda_handler'
 
 require_relative 'action_dispatch/session/dynamodb_store'
 require_relative 'active_job/queue_adapters/amazon_sqs_adapter'
+require_relative 'active_job/queue_adapters/amazon_sqs_async_adapter'
 
 require_relative 'generators/aws_record/base'