+++ lib/aws/rails/sqs_active_job/configuration.rb
@@ -30,7 +30,7 @@ module Aws
 
         # @api private
         attr_accessor :queues, :max_messages, :visibility_timeout,
-                      :shutdown_timeout, :client, :logger
+                      :shutdown_timeout, :client, :logger, :async_queue_error_handler
 
         # Don't use this method directly: Confugration is a singleton class, use
         # +Aws::Rails::SqsActiveJob.config+ to access the singleton config.
@@ -63,6 +63,12 @@ module Aws
         #   Override file to load configuration from.  If not specified will
         #   attempt to load from config/aws_sqs_active_job.yml.
         #
+        # @option options [Callable] :async_queue_error_handler An error handler
+        #   to be called when the async active job adapter experiances an error
+        #   queueing a job.  Only applies when
+        #   +active_job.queue_adapter = :amazon_sqs_async+.  Called with:
+        #   [error, job, job_options]
+        #
         # @option options [SQS::Client] :client SQS Client to use.  A default
         #   client will be created if none is provided.
         def initialize(options = {})
@@ -74,7 +80,7 @@ module Aws
         end
 
         def client
-          @client ||= Aws::SQS::Client.new
+          @client ||= Aws::SQS::Client.new(user_agent_suffix: user_agent)
         end
 
         # Return the queue_url for a given job_queue name
@@ -135,6 +141,10 @@ module Aws
         # @return [String] Configuration path found in environment or YAML file.
         def config_file_path(options)
           options[:config_file] || ENV["AWS_SQS_ACTIVE_JOB_CONFIG_FILE"]
+        end
+
+        def user_agent
+          "ft/aws-sdk-rails-activejob/#{Aws::Rails::VERSION}"
         end
       end
     end